searchdemo/requirements.txt
django
gunicorn
requests
pyelasticsearch
django-haystack
